add comment to color and size
