+++ src/components/Button/Button.tsx
@@ -85,7 +85,7 @@ const Button = ({
     <>
       {isLoading && (
         <div className="buttonLoadingIndicator">
-          <svg className="spin" alt="Loading" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16">
+          <svg className="spin" xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 16 16">
             <title>circle</title>
             <g fill="#ffffff">
               <path
