+++ src/components/Heading/Heading.stories.mdx
@@ -4,7 +4,7 @@ import { action } from '@storybook/addon-actions';
 import { withA11y } from '@storybook/addon-a11y';
 import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks';
 import { PALMETTO_FONT_SIZE_OPTIONS, PALMETTO_BRAND_COLOR_OPTIONS } from '../../lib/tokens';
-import { HEADING_LEVELS } from './Heading.constants';
+import { HEADING_LEVELS } from './Heading';
 
 <Meta
   title="Components/Heading"
