+++ types/index.d.ts
@@ -1,2 +1,4 @@
 declare module '*.scss';
 declare module '*.svg';
+declare module '@palmetto/palmetto-design-tokens/build/js/variables-size';
+declare module '@palmetto/palmetto-design-tokens/build/js/variables-color';
